improved handling of negative ArrayBuffer size fixes #708

+++ src/org/mozilla/javascript/typedarrays/NativeFloat64Array.java
@@ -36,7 +36,7 @@ public class NativeFloat64Array
 
     public NativeFloat64Array(int len)
     {
-        this(new NativeArrayBuffer(len * BYTES_PER_ELEMENT), 0, len);
+        this(new NativeArrayBuffer((double)len * BYTES_PER_ELEMENT), 0, len);
     }
 
     @Override
