+++ src/org/mozilla/javascript/typedarrays/NativeInt32Array.java
@@ -36,7 +36,7 @@ public class NativeInt32Array
 
     public NativeInt32Array(int len)
     {
-        this(new NativeArrayBuffer(len * BYTES_PER_ELEMENT), 0, len);
+        this(new NativeArrayBuffer((double)len * BYTES_PER_ELEMENT), 0, len);
     }
 
     @Override
