+++ src/org/mozilla/javascript/typedarrays/NativeUint32Array.java
@@ -35,7 +35,7 @@ public class NativeUint32Array
 
     public NativeUint32Array(int len)
     {
-        this(new NativeArrayBuffer(len * BYTES_PER_ELEMENT), 0, len);
+        this(new NativeArrayBuffer((double)len * BYTES_PER_ELEMENT), 0, len);
     }
 
     @Override
